rendering system fixed

+++ headers/utils/Math.hpp
@@ -36,8 +36,8 @@ namespace Math
 
 	inline Vec2 polar_to_cartesian(const Vec2 center, const double radius, const double angle)
 	{
-		const size_t y = static_cast<size_t>(std::round(center.y - radius * std::sin(angle)));
-		const size_t x = static_cast<size_t>(std::round(center.x + radius * std::cos(angle)));
+		const int y = static_cast<int>(std::round(center.y - radius * std::sin(angle)));
+		const int x = static_cast<int>(std::round(center.x + radius * std::cos(angle)));
 		return Vec2(y, x);
 	}
 }
